requirements.txt
streamlit
pandas
numpy
joblib
Pillow
scikit-learn 